DetailsList: Fixes
Rendering fixes for the DetailsList
