+++ src/components/DetailsList/presets/0-default.jsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import DetailsList from '../DetailsList';
+import DetailsList from '../DetailsList_test';
 
 
 const searchIcon = "Filter";
@@ -18,9 +18,9 @@ Column B
 Column C
 Column D
 Actions`;
-const defaultRows = `link(UXPin.com|www.uxpin.com), icon(CircleCheckSolid|success) Ready, C-1, D-1, icon(Deploy|themePrimary) Deploy
-link(Component_Name_B), icon(CircleWarningSolid|warning) Restarting..., C-2, D-2, icon(MoreVertical|themePrimary)
-link(Component_Name_C), icon(CircleClearSolid|error) Unavailable, C-3, D-3, icon(MoreVertical|themePrimary)`;
+const defaultRows = `link(UXPin.com|www.uxpin.com), icon(SkypeCircleCheck|success) Ready, C-1, D-1, icon(Deploy|themePrimary) Deploy
+link(Component_Name_B), icon(WarningSolid|warning) Restarting..., C-2, D-2, icon(MoreVertical|themePrimary)
+link(Component_Name_C), icon(StatusErrorFull|error) Unavailable, C-3, D-3, icon(MoreVertical|themePrimary)`;
